+++ R/fipio.R
@@ -1,3 +1,51 @@
+#' @title Convert a state name, abbreviation, or county name to FIPS codes
+#' @param state State names or abbreviations
+#' @param county County names
+#' @return a `character` vector
+#' @examples
+#' fipio::as_fips(state = "California")
+#' fipio::as_fips(state = "NC")
+#' fipio::as_fips(state = "Rhode Island", county = "Washington")
+#' fipio::as_fips(c("CA", "North Carolina"), c("Stanislaus", "NEW HANOVER"))
+#'
+#' @export
+as_fips <- function(state, county = NULL) {
+    if (missing(state) | any(state == "") | is.null(state)) {
+        stop("`state` must be specificed at least.", call. = FALSE)
+    }
+
+    state   <- tolower(state)
+    indices <- match(state, tolower(fips_$state_abbr))
+
+    tmp <- match(state, tolower(fips_$state_name))
+    indices[is.na(indices)] <- tmp[!is.na(tmp)]
+    rm(tmp)
+
+    ret <- fips_$state_code[indices]
+
+    if (any(!is.null(county))) {
+        state_abbrs  <- fips_$state_abbr[indices]
+        county       <- trimws(gsub("county", "", tolower(county)))
+        county_tbl   <- fips_[fips_$state_abbr %in% state_abbrs, c(3, 4, 6)]
+        county_codes <- county_tbl$fip_code[
+            match(county, tolower(county_tbl$county_name))
+        ]
+
+        if (all(is.na(county_codes))) {
+            repl <- TRUE
+        } else {
+            repl <- !is.na(c(
+                county_codes,
+                rep(NA, length(ret) - length(county_codes))
+            ))
+        }
+
+        ret[repl] <- county_codes[repl]
+    }
+
+    ret
+}
+
 #' @title Get the state abbreviation for a FIPS code
 #' @param fip 2-digit or 5-digit FIPS code
 #' @return a `character` vector
@@ -96,7 +144,19 @@ fips_metadata <- function(fip, geometry = FALSE) {
     df
 }
 
-#' @keywords internal
-.has_sfheaders <- function() {
-    requireNamespace("sfheaders", quietly = TRUE)
+#nocov start
+#' @title Get the matching function that `fipio` is using
+#' @description
+#' This function is primarily for debugging purposes,
+#' or for ensuring that the correct matching function
+#' is used.
+#' @return `TRUE` if `fastmatch::fmatch` is used.
+#' @export
+using_fastmatch <- function() {
+    if (getNamespaceName(environment(match))[[1]] == "fastmatch") {
+        TRUE
+    } else {
+        FALSE
+    }
 }
+#nocov end